module3/after/model_deploy_aci/model_requirements.txt
nltk
numpy
onnxruntime
azureml-core
azureml-defaults